+++ cli/bench/report.cc
@@ -27,9 +27,9 @@
 #include <vlink/base/format.h>
 #include <vlink/base/helpers.h>
 #include <vlink/base/message_loop.h>
+#include <vlink/base/terminal_stream.h>
 #include <vlink/base/timer.h>
 #include <vlink/base/utils.h>
-#include <vlink/extension/terminal_stream.h>
 
 #include <algorithm>
 #include <array>