+++ cli/monitor/monitor.cc
@@ -23,9 +23,9 @@
 
 #include <vlink/base/elapsed_timer.h>
 #include <vlink/base/helpers.h>
+#include <vlink/base/terminal_stream.h>
 #include <vlink/base/utils.h>
 #include <vlink/extension/discovery_viewer.h>
-#include <vlink/extension/terminal_stream.h>
 #include <vlink/version.h>
 #include <vlink/vlink.h>